feat: docker compose + swagger + api endpoints

+++ src/main/java/com/arodriguez/subscriptionmanager/repository/SubscriptionRepository.java
@@ -4,11 +4,13 @@ import com.arodriguez.subscriptionmanager.entity.Subscription;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
+
 public interface SubscriptionRepository extends JpaRepository<Subscription, Long> {
 
     @Query("select coalesce(sum(s.monthlyCost), 0) from Subscription s")
-    Double sumTotalMonthlyCost();
+    BigDecimal sumTotalMonthlyCost();
 
     @Query("select coalesce(sum(s.monthlyCost), 0) from Subscription s where s.active = true")
-    Double sumActiveMonthlyCost();
+    BigDecimal sumActiveMonthlyCost();
 }
